+++ packages/react-components/react-avatar-context/src/index.ts
@@ -1,2 +1,0 @@
-// TODO: replace with real exports
-export {};