+++ packages/react-components/react-combobox/src/components/Listbox/renderListbox.tsx
@@ -1,5 +1,9 @@
-import * as React from 'react';
-import { getSlots } from '@fluentui/react-utilities';
+/** @jsxRuntime classic */
+/** @jsx createElement */
+
+import { createElement } from '@fluentui/react-jsx-runtime';
+
+import { getSlotsNext } from '@fluentui/react-utilities';
 import type { ListboxContextValues, ListboxState, ListboxSlots } from './Listbox.types';
 import { ListboxContext } from '../../contexts/ListboxContext';
 
@@ -7,7 +11,7 @@ import { ListboxContext } from '../../contexts/ListboxContext';
  * Render the final JSX of Listbox
  */
 export const renderListbox_unstable = (state: ListboxState, contextValues: ListboxContextValues) => {
-  const { slots, slotProps } = getSlots<ListboxSlots>(state);
+  const { slots, slotProps } = getSlotsNext<ListboxSlots>(state);
 
   return (
     <ListboxContext.Provider value={contextValues.listbox}>