+++ packages/react-components/react-drawer/stories/Drawer/index.stories.tsx
@@ -1,9 +1,17 @@
 import { Drawer } from '@fluentui/react-drawer';
 import descriptionMd from './DrawerDescription.md';
-import DrawerBestpracticesMd from './DrawerBestPractices.md';
+import bestPracticesMd from './DrawerBestPractices.md';
 import previewMd from './DrawerPreview.md';
 
 export { Default } from './DrawerDefault.stories';
+export { DefaultOpen } from './DrawerDefaultOpen.stories';
+export { AlwaysOpen } from './DrawerAlwaysOpen.stories';
+export { PreventClose } from './DrawerPreventClose.stories';
+export { Position } from './DrawerPosition.stories';
+export { Inline } from './DrawerInline.stories';
+export { Size } from './DrawerSize.stories';
+export { CustomSize } from './DrawerCustomSize.stories';
+export { Separator } from './DrawerSeparator.stories';
 
 export default {
   title: 'Preview Components/Drawer/Drawer',
@@ -11,7 +19,7 @@ export default {
   parameters: {
     docs: {
       description: {
-        component: [descriptionMd, DrawerBestpracticesMd, previewMd].join('\n'),
+        component: [descriptionMd, bestPracticesMd, previewMd].join('\n'),
       },
     },
   },