+++ packages/react-components/react-field/src/components/Field/Field.test.tsx
@@ -1,7 +1,20 @@
 import * as React from 'react';
+
+import { Tooltip } from '@fluentui/react-tooltip';
 import { render } from '@testing-library/react';
+import { useFieldControlProps_unstable } from '../../contexts/useFieldControlProps';
 import { isConformant } from '../../testing/isConformant';
 import { Field } from './index';
+
+const TestInput = React.forwardRef<HTMLInputElement, React.InputHTMLAttributes<HTMLInputElement>>((props, ref) => {
+  props = useFieldControlProps_unstable(props, { supportsLabelFor: true, supportsRequired: true });
+  return <input ref={ref} {...props} />;
+});
+
+const TestGroup = React.forwardRef<HTMLDivElement, React.HTMLAttributes<HTMLDivElement>>((props, ref) => {
+  props = useFieldControlProps_unstable(props);
+  return <div role="group" ref={ref} {...props} />;
+});
 
 describe('Field', () => {
   isConformant({
@@ -21,24 +34,10 @@ describe('Field', () => {
     },
   });
 
-  it("sets the label's htmlFor to the child's id if it has one", () => {
+  it('generates an id for the control and uses it as the label.htmlFor', () => {
     const result = render(
       <Field label="Test label">
-        <input id="test-id" />
-      </Field>,
-    );
-
-    const input = result.getByRole('textbox');
-    const label = result.getByText('Test label') as HTMLLabelElement;
-
-    expect(input.id).toBe('test-id');
-    expect(label.htmlFor).toBe('test-id');
-  });
-
-  it('generates an id for the child if it does not have one,', () => {
-    const result = render(
-      <Field label="Test label">
-        <input />
+        <TestInput />
       </Field>,
     );
 
@@ -47,38 +46,92 @@ describe('Field', () => {
 
     expect(input.id).toBeTruthy();
     expect(label.htmlFor).toBe(input.id);
+    expect(input.getAttribute('aria-labelledby')).toBeFalsy();
   });
 
-  it('sets aria-labelledby on the control', () => {
+  it('does not set aria-labelledby if label.htmlFor matches the control id', () => {
     const result = render(
-      <Field label="Test label">
-        <input />
+      <Field label={{ children: 'Test label', htmlFor: 'test-label-for' }}>
+        <TestInput id="test-label-for" />
       </Field>,
     );
+
+    const input = result.getByRole('textbox');
+    const label = result.getByText('Test label') as HTMLLabelElement;
+
+    expect(label.htmlFor).toBe('test-label-for');
+    expect(input.id).toBe('test-label-for');
+    expect(input.getAttribute('aria-labelledby')).toBeFalsy();
+  });
+
+  it('falls back to aria-labelledby if the control has an id that does not match the label.htmlFor', () => {
+    const result = render(
+      <Field label="Test label">
+        <TestInput id="test-id" /> {/* Does not match label's generated htmlFor */}
+      </Field>,
+    );
+
     const input = result.getByRole('textbox');
     const label = result.getByText('Test label') as HTMLLabelElement;
 
     expect(label.id).toBeTruthy();
     expect(input.getAttribute('aria-labelledby')).toBe(label.id);
+    expect(input.id).toBe('test-id');
   });
 
-  it('adds a required asterisk * to the label, and aria-required on the control when required is set', () => {
+  it('sets aria-labelledby on a control that does not support label.htmlFor', () => {
+    const result = render(
+      <Field label="Test label">
+        <TestGroup /> {/* Groups do not support label.htmlFor */}
+      </Field>,
+    );
+    const group = result.getByRole('group');
+    const label = result.getByText('Test label') as HTMLLabelElement;
+
+    expect(label.id).toBeTruthy();
+    expect(group.getAttribute('aria-labelledby')).toBe(label.id);
+  });
+
+  it('adds a required asterisk * to the label when required is set', () => {
     const result = render(
       <Field label="Test label" required>
-        <input />
+        <TestInput />
       </Field>,
     );
 
-    const input = result.getByRole('textbox');
-
     expect(result.getByText('*')).toBeTruthy();
-    expect(input.getAttribute('aria-required')).toBe('true');
+  });
+
+  it('sets `required` on a control that supports the `required` prop', () => {
+    const result = render(
+      <Field required>
+        <TestInput />
+      </Field>,
+    );
+
+    const input = result.getByRole('textbox') as HTMLInputElement;
+
+    expect(input.required).toBe(true);
+    expect(input.getAttribute('aria-required')).toBeNull();
+  });
+
+  it('sets `aria-required` on a control that does not support the `required` prop', () => {
+    const result = render(
+      <Field required>
+        <TestGroup /> {/* Groups do not support the required prop */}
+      </Field>,
+    );
+
+    const group = result.getByRole('group');
+
+    expect(group.getAttribute('aria-required')).toBe('true');
+    expect(group.getAttribute('required')).toBeNull();
   });
 
   it('sets aria-describedby to the hint', () => {
     const result = render(
       <Field hint="Test hint">
-        <input />
+        <TestInput />
       </Field>,
     );
     const input = result.getByRole('textbox');
@@ -91,7 +144,7 @@ describe('Field', () => {
   it('sets aria-describedby to the validationMessage', () => {
     const result = render(
       <Field validationMessage="Test validation message" validationState="warning">
-        <input />
+        <TestInput />
       </Field>,
     );
     const input = result.getByRole('textbox');
@@ -104,7 +157,7 @@ describe('Field', () => {
   it('sets aria-describedby to the validationMessage + hint', () => {
     const result = render(
       <Field hint="Test hint" validationMessage="Test validation message">
-        <input />
+        <TestInput />
       </Field>,
     );
     const input = result.getByRole('textbox');
@@ -117,7 +170,7 @@ describe('Field', () => {
   it('sets aria-describedby to the validationMessage + hint + user aria-describedby', () => {
     const result = render(
       <Field hint="Test hint" validationMessage="Test validation message">
-        <input aria-describedby="test-describedby" />
+        <TestInput aria-describedby="test-describedby" />
       </Field>,
     );
     const input = result.getByRole('textbox');
@@ -130,7 +183,7 @@ describe('Field', () => {
   it('sets aria-invalid if an error', () => {
     const result = render(
       <Field validationState="error">
-        <input />
+        <TestInput />
       </Field>,
     );
     const input = result.getByRole('textbox');
@@ -138,18 +191,61 @@ describe('Field', () => {
     expect(input.getAttribute('aria-invalid')).toBeTruthy();
   });
 
-  it('does not override user aria props, EXCEPT aria-describedby', () => {
+  it('does not override user props (other than aria-describedby)', () => {
     const result = render(
-      <Field label="test label" validationMessage="test description" hint="test hint">
-        <input aria-labelledby="test-labelledby" aria-errormessage="test-errormessage" aria-invalid={false} />
+      <Field label="test label" validationMessage="test description" hint="test hint" required>
+        <TestInput
+          id="test-id"
+          aria-labelledby="test-labelledby"
+          aria-errormessage="test-errormessage"
+          aria-invalid={false}
+          aria-required={false}
+        />
       </Field>,
     );
 
     const input = result.getByRole('textbox');
 
+    expect(input.id).toBe('test-id');
     expect(input.getAttribute('aria-labelledby')).toBe('test-labelledby');
     expect(input.getAttribute('aria-errormessage')).toBe('test-errormessage');
     expect(input.getAttribute('aria-invalid')).toBe('false');
+    expect(input.getAttribute('aria-required')).toBe('false');
+    // aria-describedby gets merged with the hint and validationMessage; that is tested above
+  });
+
+  it('passes props through other component(s) using context', () => {
+    const result = render(
+      <Field label="Test label" hint="test hint" required>
+        <div>
+          <span>...</span>
+          <TestInput />
+        </div>
+      </Field>,
+    );
+
+    const input = result.getByRole('textbox');
+    const label = result.getByText('Test label') as HTMLLabelElement;
+    const hint = result.getByText('test hint');
+
+    expect(label.htmlFor).toBe(input.id);
+    expect(input.getAttribute('aria-describedby')).toBe(hint.id);
+  });
+
+  it('merges Field describedby with Tooltip describedby', () => {
+    const result = render(
+      <Field hint="Test hint">
+        <Tooltip relationship="description" content="Test tooltip">
+          <TestInput />
+        </Tooltip>
+      </Field>,
+    );
+
+    const input = result.getByRole('textbox');
+    const hint = result.getByText('Test hint');
+    const tooltip = result.getByText('Test tooltip');
+
+    expect(input.getAttribute('aria-describedby')).toBe(`${hint.id} ${tooltip.id}`);
   });
 
   it.each([
@@ -161,7 +257,7 @@ describe('Field', () => {
   ] as const)('if validationState is %s, sets role to %s on the validationMessage', (validationState, role) => {
     const result = render(
       <Field validationState={validationState} validationMessage="test validation message">
-        <input />
+        <TestInput />
       </Field>,
     );
     const validationMessage = result.getByText('test validation message');