+++ packages/react-components/react-menu/src/components/MenuSplitGroup/MenuSplitGroup.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { useMenuSplitGroup_unstable } from './useMenuSplitGroup';
 import { renderMenuSplitGroup_unstable } from './renderMenuSplitGroup';
-import { useMenuSplitGroupStyles_unstable } from './useMenuSplitGroupStyles';
+import { useMenuSplitGroupStyles_unstable } from './useMenuSplitGroupStyles.styles';
 import type { MenuSplitGroupProps } from './MenuSplitGroup.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
-import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Layout wrapper that provides extra keyboard navigation behavior for two `MenuItem` components.
@@ -14,8 +14,7 @@ export const MenuSplitGroup: ForwardRefComponent<MenuSplitGroupProps> = React.fo
 
   useMenuSplitGroupStyles_unstable(state);
 
-  const { useMenuSplitGroupStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
-  useCustomStyles(state);
+  useCustomStyleHook_unstable('useMenuSplitGroupStyles_unstable')(state);
 
   return renderMenuSplitGroup_unstable(state);
 });