+++ packages/react-components/react-switch/src/components/Switch/Switch.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { useSwitch_unstable } from './useSwitch';
 import { renderSwitch_unstable } from './renderSwitch';
-import { useSwitchStyles_unstable } from './useSwitchStyles';
+import { useSwitchStyles_unstable } from './useSwitchStyles.styles';
 import type { SwitchProps } from './Switch.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
-import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * Switches enable users to trigger an option on or off through pressing the component.
@@ -14,8 +14,7 @@ export const Switch: ForwardRefComponent<SwitchProps> = React.forwardRef((props,
 
   useSwitchStyles_unstable(state);
 
-  const { useSwitchStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
-  useCustomStyles(state);
+  useCustomStyleHook_unstable('useSwitchStyles_unstable')(state);
 
   return renderSwitch_unstable(state);
 });