+++ packages/react-components/react-table/src/components/TableCellActions/renderTableCellActions.tsx
@@ -1,12 +1,15 @@
-import * as React from 'react';
-import { getSlots } from '@fluentui/react-utilities';
+/** @jsxRuntime classic */
+/** @jsx createElement */
+
+import { createElement } from '@fluentui/react-jsx-runtime';
+import { getSlotsNext } from '@fluentui/react-utilities';
 import type { TableCellActionsState, TableCellActionsSlots } from './TableCellActions.types';
 
 /**
  * Render the final JSX of TableCellActions
  */
 export const renderTableCellActions_unstable = (state: TableCellActionsState) => {
-  const { slots, slotProps } = getSlots<TableCellActionsSlots>(state);
+  const { slots, slotProps } = getSlotsNext<TableCellActionsSlots>(state);
 
   return <slots.root {...slotProps.root} />;
 };