+++ packages/react-components/react-table/src/components/TableRow/index.ts
@@ -2,4 +2,4 @@ export * from './TableRow';
 export * from './TableRow.types';
 export * from './renderTableRow';
 export * from './useTableRow';
-export * from './useTableRowStyles';
+export * from './useTableRowStyles.styles';