+++ packages/react-components/react-text/src/components/presets/Subtitle2Stronger/index.ts
@@ -1,1 +1,2 @@
 export * from './Subtitle2Stronger';
+export { subtitle2StrongerClassNames } from './useSubtitle2Stronger.styles';