+++ scripts/storybook/src/utils.js
@@ -3,7 +3,7 @@ const path = require('path');
 
 const { fullSourcePlugin: babelPlugin } = require('@fluentui/babel-preset-storybook-full-source');
 const { isConvergedPackage, getAllPackageInfo, getProjectMetadata } = require('@fluentui/scripts-monorepo');
-const { stripIndents, offsetFromRoot, workspaceRoot, readJsonFile, writeJsonFile } = require('@nrwl/devkit');
+const { stripIndents, offsetFromRoot, workspaceRoot } = require('@nrwl/devkit');
 const semver = require('semver');
 const { TsconfigPathsPlugin } = require('tsconfig-paths-webpack-plugin');
 
@@ -106,7 +106,7 @@ function loadWorkspaceAddon(addonName, options) {
     const { registerTsPaths } = require('@fluentui/scripts-storybook');
 
     function managerWebpack(config, options) {
-      registerTsPaths({config, tsConfigPath: '${posixTsConfigPath}'});
+      registerTsPaths({config, configFile: '${posixTsConfigPath}'});
       return config;
     }
 
@@ -129,6 +129,8 @@ function loadWorkspaceAddon(addonName, options) {
  * @returns {import("webpack").RuleSetRule}
  */
 function _createCodesandboxRule(allPackageInfo = getAllPackageInfo()) {
+  const config = getCodesandboxBabelOptions();
+
   return {
     /**
      * why the usage of 'post' ? - we need to run this loader after all storybook webpack rules/loaders have been executed.
@@ -141,7 +143,7 @@ function _createCodesandboxRule(allPackageInfo = getAllPackageInfo()) {
     use: {
       loader: 'babel-loader',
       options: _processBabelLoaderOptions({
-        plugins: [[babelPlugin, getCodesandboxBabelOptions()]],
+        plugins: [[babelPlugin, config]],
       }),
     },
   };
@@ -150,36 +152,48 @@ function _createCodesandboxRule(allPackageInfo = getAllPackageInfo()) {
    * @returns {import('@fluentui/babel-preset-storybook-full-source').BabelPluginOptions}
    */
   function getCodesandboxBabelOptions() {
+    /**
+     * packages that are part of v9 but are not meant for platform:web
+     */
+    const excludePackages = [
+      '@fluentui/babel-preset-storybook-full-source',
+      '@fluentui/react-storybook-addon',
+      '@fluentui/react-storybook-addon-codesandbox',
+      '@fluentui/react-conformance-griffel',
+    ];
+
+    // TODO: https://github.com/microsoft/fluentui/issues/26691
+    const packagesOutsideReactComponentsSuite = [
+      '@fluentui/react-data-grid-react-window',
+      '@fluentui/react-datepicker-compat',
+      '@fluentui/react-migration-v8-v9',
+      '@fluentui/react-migration-v0-v9',
+    ];
+
     const importMappings = Object.values(allPackageInfo).reduce((acc, cur) => {
+      if (excludePackages.includes(cur.packageJson.name)) {
+        return acc;
+      }
+
+      if (packagesOutsideReactComponentsSuite.includes(cur.packageJson.name)) {
+        acc[cur.packageJson.name] = { replace: cur.packageJson.name };
+        return acc;
+      }
+
       if (isConvergedPackage({ packagePathOrJson: cur.packageJson, projectType: 'library' })) {
         const isPrerelease = semver.prerelease(cur.packageJson.version) !== null;
 
         acc[cur.packageJson.name] = isPrerelease
           ? { replace: '@fluentui/react-components/unstable' }
           : { replace: '@fluentui/react-components' };
+
+        return acc;
       }
 
       return acc;
     }, /** @type import('@fluentui/babel-preset-storybook-full-source').BabelPluginOptions*/ ({}));
 
-    return {
-      ...importMappings,
-
-      // TODO: https://github.com/microsoft/fluentui/issues/26691
-
-      '@fluentui/react-data-grid-react-window': {
-        replace: '@fluentui/react-data-grid-react-window',
-      },
-      '@fluentui/react-migration-v8-v9': {
-        replace: '@fluentui/react-migration-v8-v9',
-      },
-      '@fluentui/react-migration-v0-v9': {
-        replace: '@fluentui/react-migration-v0-v9',
-      },
-      '@fluentui/react-datepicker-compat': {
-        replace: '@fluentui/react-datepicker-compat',
-      },
-    };
+    return importMappings;
   }
 }
 
@@ -340,48 +354,9 @@ function overrideDefaultBabelLoader(options) {
   }
 }
 
-/**
- * Create tsconfig.json with merged "compilerOptions.paths" from v0,v8,v9 tsconfigs.
- *
- * Main purpose of this is to be used for build-less DX in webpack in tandem with {@link registerTsPaths}
- * @returns
- */
-function createPathAliasesConfig() {
-  const { tsConfigAllPath } = createMergedTsConfig();
-  return { tsConfigAllPath };
-}
-
-function createMergedTsConfig() {
-  const rootPath = workspaceRoot;
-  const tsConfigAllPath = path.join(rootPath, 'dist/tsconfig.base.all.json');
-  const baseConfigs = {
-    v0: readJsonFile(path.join(rootPath, 'tsconfig.base.v0.json')),
-    v8: readJsonFile(path.join(rootPath, 'tsconfig.base.v8.json')),
-    v9: readJsonFile(path.join(rootPath, 'tsconfig.base.json')),
-  };
-  const mergedTsConfig = {
-    compilerOptions: {
-      moduleResolution: 'node',
-      forceConsistentCasingInFileNames: true,
-      skipLibCheck: true,
-      baseUrl: workspaceRoot,
-      paths: {
-        ...baseConfigs.v0.compilerOptions.paths,
-        ...baseConfigs.v8.compilerOptions.paths,
-        ...baseConfigs.v9.compilerOptions.paths,
-      },
-    },
-  };
-
-  writeJsonFile(tsConfigAllPath, mergedTsConfig);
-
-  return { tsConfigAllPath, mergedTsConfig };
-}
-
 exports.getPackageStoriesGlob = getPackageStoriesGlob;
 exports.loadWorkspaceAddon = loadWorkspaceAddon;
 exports.registerTsPaths = registerTsPaths;
 exports.registerRules = registerRules;
-exports.createPathAliasesConfig = createPathAliasesConfig;
 exports.overrideDefaultBabelLoader = overrideDefaultBabelLoader;
 exports._createCodesandboxRule = _createCodesandboxRule;